requirements.txt
matplotlib==3.3.4
numpy==1.19.2
PyYAML==6.0
torch==1.9.0
torchvision==0.10.0
